# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bokks Library Api",
    "description": " API contracts for books Library ",
    "version": "1"
  },
  "paths": {
...
    "/api/Books": {
      "get": {
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookQuery"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "GetBooks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Books/api/books/filterBooks": {
      "post": {
        "tags": [
          "Books"
        ],
        "operationId": "FilterBooks",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Books/api/books/filterOptions": {
      "post": {
        "tags": [
          "Books"
        ],
        "operationId": "GetFilters",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilterQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptions"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptions"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptions"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Author": {
        "required": [
          "Address",
          "Email",
          "FirstName",
          "LastName",
          "PhoneNumber"
        ],
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "PhoneNumber": {
            "type": "string"
          },
          "Address": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "BookAuthors": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "BookId": {
            "type": "integer",
            "format": "int32"
          },
          "Book": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Book"
              }
            ],
            "nullable": true
          },
          "AuthorId": {
            "type": "integer",
            "format": "int32"
          },
          "Author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Editor": {
        "required": [
          "Address",
          "Email",
          "FirstName",
          "LastName",
          "PhoneNumber"
        ],
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "PhoneNumber": {
            "type": "string"
          },
          "Address": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Tag": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookTag": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "BookId": {
            "type": "integer",
            "format": "int32"
          },
          "Book": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Book"
              }
            ],
            "nullable": true
          },
          "TagId": {
            "type": "integer",
            "format": "int32"
          },
          "Tag": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Tag"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Comment": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RessourceFormat": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Ressource": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Format": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RessourceFormat"
              }
            ]
          },
          "Location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Category": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "subCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategory"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SubCategory": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "CategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "Category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookSubCategies": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "BookId": {
            "type": "integer",
            "format": "int32"
          },
          "Book": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Book"
              }
            ],
            "nullable": true
          },
          "SubCategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "SubCategory": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SubCategory"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Book": {
        "required": [
          "AuthorLinks",
          "Comment",
          "Editor",
          "ImageLocation",
          "Name",
          "Ressource",
          "SubCategoriesLink",
          "TagLinks"
        ],
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string"
          },
          "ImageLocation": {
            "type": "string"
          },
          "AuthorLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookAuthors"
            }
          },
          "Editor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Editor"
              }
            ]
          },
          "TagLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookTag"
            }
          },
          "Comment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "Ressource": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Ressource"
              }
            ]
          },
          "SubCategoriesLink": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookSubCategies"
            }
          }
        },
        "additionalProperties": false
      },
      "CreateBookQuery": {
        "required": [
          "Authors",
          "EditorId",
          "ImageLocation",
          "Name",
          "Ressource",
          "SubCategories",
          "Tags"
        ],
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          },
          "ImageLocation": {
            "type": "string"
          },
          "Authors": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "EditorId": {
            "type": "integer",
            "format": "int32"
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "Ressource": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Ressource"
              }
            ]
          },
          "SubCategories": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "additionalProperties": false
      },
      "FilterQuery": {
        "type": "object",
        "properties": {
          "Authors": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "Editors": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "Categories": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "SubCategories": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookResponse": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "ImageLocation": {
            "type": "string",
            "nullable": true
          },
          "Authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "nullable": true
          },
          "Editor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Editor"
              }
            ],
            "nullable": true
          },
          "Tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          },
          "Comment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "nullable": true
          },
          "Ressource": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Ressource"
              }
            ],
            "nullable": true
          },
          "SubCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategory"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FilterOptions": {
        "type": "object",
        "properties": {
          "Authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "nullable": true
          },
          "Editors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Editor"
            },
            "nullable": true
          },
          "Categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "nullable": true
          },
          "SubCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategory"
            },
            "nullable": true
          },
          "Tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}